Split middleware into separate class
